HH-30: изменяет Mapper у Search и Favourite
